requirements.txt
matplotlib==3.10.0
pandas==2.2.3
protobuf==5.29.3
scikit_learn==1.6.1
seaborn==0.13.2
streamlit==1.40.1
google-generativeai==0.8.4